don't send [create_game]
previously when we started a new game we first sended [create_game] and
then the contents of the level. the [create_game] contained no
information that was not in the level date except the pasword. Now we
don't send [start_game] anymore, and sinstead send the password with the
level data. This simplyfies the network game setup, especialy because we
don't have to think about the 'between [create_game] and level
initilisation' state on the wesnothd::game objects anymore. Also we move
some tweaks of [multiplayer]require_ era/modification/scenario to the
client. The main reason why that is easier is just that the config class
is simpler to handle than the simple_wml class.