make force_lock_settings default to true for [scenario]
In most cases the campaigns designer don't want the use to be able to
overwrite scenario properites like "turn limit" or "use fog".

It is easiert to do this here in the code that loads the config than in
the code that reads that value becasue there are multiple codes that
read it and it's not clear that all of then know whether teh scenario
was a [multiplayer] or a [scenario].

Partly implements http://gna.org/bugs/?23037 which suggests to remove
the ability to overwite those settings in capaigns completely.