fixup 2486b8ba4bcef056fb523d91f4a8445b48aef685
That commit overlooked the tstring_concat_aux function in
lua_common.cpp, a helper function for tstring concatenation. This
caused lua errors when HttT was started, arising in the code to
display objectives.

This commit fixes up the tstring_concat_aux function to more
directly check the tstring type.

It also fixes luaW_totstring and luaW_toconfig so that they don't
immediately flag a type error if they fail to convert userdata to
a tstring... this was caused by inappropriately using
luaL_checkudata when luaL_testudata should be used.