partially fix 23076 (don't hold std::vector<unit> in recall dialog)
After this commit 9e7dc5ba008298d823c13cff214f7477601b5efd which
introduced reference counted unit pointers, the recall dialog
began to fail assertions. The problem was that it was making
private copies of all units, and when units get deleted this list
got out of sync. I'm not totally sure exactly what mechanism was
making it work before, but this commit which shares the units list
between the different gui elements prevents the assertion failure.

Note that there is still a bug, in that the unit preview pane at
left still gets out of sync somehow...