fully fix bug #23076
This commit keeps the various UI elements of the recall, unit list
dialogs in sync by passing a smart pointer to the underlying set
of units between them. This seems to fix all of the problems
mentioned in 23076, preventing assertion failures and mismatches
between the unit list and the preview pane.