T S1: reduce chance of quintain for dying before turn 4
The reason for this is that if they player is lucky, they might kill him before turn 4,
and when it dies, Delf leaves, preventing the entire attack exposition from happening.
This cuts off the student's moves after they move to keep on turn 3, and adds a dialog line
by Delf regarding it.

This doesn't fix any situation where the player disregards instructions and doesn't return to
the keep, but that's their problem.