lua: fixup 9eaba64651944bbe463a77a2611f37697421f15d
This commit broke the lua unit modify method for unit.side, the
reason was that I wasn't sure how to handle a safety check that
required a reference to the gamestate -- I'm not totally convinced
that units should not be allowed to be made in lua if there is not
an active gamestate.

However instead of commenting out this line, I should have
replaced it with a version that does the same thing without the
safety check... not sure why I didn't do this the first time.

This fixes lua errors that occur in the test scenario startup.