don't do variable subsitution when parsing [team][village]
Becasue are no posibilities to change the variables or the positions of
the villages before this code happens, i cannot see any common usecase
for this feature, Also it's not documented in the wiki.

Note that it is of course still possible to change village owners
depending on variables in a prestart event.