cosmetic: tidy up punctuation in controller ctor initializer lists
Moves the : and , to lead the lines rather than end them. This
makes the initializer lists a bit more readable, as it lines up
these punctuations and separates them from the initializer
expressions. It seems to be mordante's style in most of the GUI2
code to do it this way. It also has the advantage that each line
is paired with its matching punctuation mark, so if you need to
add or rearrange things, you never need to edit a line solely to
move a comma around, reducing the noise in the diffs.