lua: change a reference variable to a non-reference type
This line was not an error -- it is legal in C++ to bind a const
reference to a temporary value. It's only slightly more confusing
than usual in this case because the string it is built from is
a C-string owned by lua, and the temporary std::string object is
implicit.

I don't believe this commit will even make any difference in the
compiled code, it's only slightly simpler and more readable.