preproc: Use the 'preprocessor' log domain for all diagnostics
We formerly used 'config', which is also used by the parser. This
resulted in a massive amount of noise being generated when trying to
obtain more details about the preprocessor's actions, since 'config' is
the go-to domain for pretty much everything involving WML (not even just
the parser). Since the introduction of --preprocess added a dedicated
log domain used by functions related to that switch, we might as well
stick to that.