campaignd: Catch utils::invalid_utf8_exception within handle_upload()
Otherwise add-ons with invalid UTF-8 sequences in the add-on name or the
fields matched against the blacklist (description, author, email, title)
crash campaignd since the exception is not handled anywhere.

We specifically want to handle this exception deep within
handle_upload() and with more specific knowledge of the circumstances
because otherwise we can't give the user an error message with enough
information to help them.