don't reload config when selecing campaigns
this reverts
https://github.com/wesnoth/wesnoth/commit/7ca092dc215810fc68017b8bef673c6edbe0f68e
and fixes http://gna.org/bugs/index.php?23229 and
http://gna.org/bugs/?23038

This means that [era]s and [modification]s inside an '#ifdef
MULTIPLAYER' might be loaded during the sp campaign selection. But
that's the smaller pain compared to the wml reload. A proper fix for
that one would be type= attribute for [era] and [modification] just like
for [campaign].