don't do auto delete save on defeat in mp campaigns
We now use 'end_level.proceed_to_next_level' instead of 'res == VICTORY
|| (game_type == game_classification::MULTIPLAYER && res == DEFEAT'  to
decide whether we do auto save deletion. The intention is to adress 'if
MP campaigns ever work again, we might need to change this test'
as it was commented in the code

We also refactor playsingle/mp_scenario to not do
playcontroller.update_savegame_snapshot(); if res == DEFEAT since it is
not neeeded in this case.