fix bug #22984: sliders not adjustable with arrow keys
This commit causes sliders to attract focus properly when their
grips are clicked on, and to respond to left and right arrow
presses again. It partly reverts two earlier commits that
affected this functionality.

cf84e1d8aae84b76e7e9adc15847089bdc940632
49614599fa5211e0bc3fdba43c6255509da66022

It causes the zoom slider to behave slightly badly, because it
does not lose focus when the user clicks elsewhere on the map.
But this bug should be fixed at the source and not by disabling
wanted functionality for all sliders.