add support for styled fonts outside of SDL TTF in GUI1, help
This commit allows bold and italic font files to be specified in
the fonts.cfg file. When attribute "bold_name" is set to the name
of a font file, this font will be used in place of SDL TTF bold-
styling applied to the base font. Similarly for "italic_name".

This commit assumes that the code points for the styled versions
are the same as for the unstyled version. If that assumption is
not true for some font that we care about then this commit will
likely have to be revised.