fix mp lobby autodownloading of addons
Makes the following changes:

When we download new add-ons,
- tear down the mp ui by throwing an exception
- refresh the changed game config, and reload for this game mode
- relaunch the mp lobby, without making a new network connection
- make sure to ask the mp lobby to refresh the games list, so that
we update the joinable / needs addon status appropriately.

After this commit, I have tested that I can prompt other clients
to download add-ons, and that they will do it successsfully and
then be able to join the game.

Still do: Version checking. And, check if the host has an old
version.