wmltools: refine comma_split function
First: minor change of the argument name passed to the function
from "value" to "csstring", since the string might not be an
attribute value.

Second: more testing pointed out that an empty list would not
meet the "if list:" condition and would not be extended. So
make it explicitly "!= None".

Third: when asking on IRC about trailing whitespace, it was
noted that utils::split also removed empty entries, which
reminded me that I should do that as well.

The biggest change is that the function will now rstrip() by
default, with other options for dealing with trailing
whitespace offered.