add addon version checking and reporting to multiplayer lobby
This also includes a refactor of the "set_game_items" function,
which was quite large and difficult to understand at a glance.
The refactor splits most of the work among 5 or 6 helper functions.

We also add some reporting of add-on status to the name, just as
with the (Unknown Era) string.

We also give the mp lobby a reference, at construction time, to
the list of currently installed add-ons, which will have the same
lifespan as the game_config reference which it is passed. This
ensures that this list is not changing based on add-on directory
contents, unless the lobby gets rebuilt. This is for the sanity
of the lobby, but it also makes testing easier.