fix an assertion failure if an add-on is not available on server
If someone hosts an mp game using an add-on which is not published,
the client of another player cannot know that, so it may prompt
them to connect to the add-on server and try to download.

However the previous implementation of "ad_hoc_addon_fetch_session"
used the "addon_at" helper function, which gives an assertion
failure if the add-on does not exist, presumably because this was
used only with the add-on manager gui before.

In this commit we change the fetch session function to explicitly
search for the addon id and handle a missing add-on with an error
message rather than assertion failure.