Remove superfluous 'sane' flag from version_info
This also removes the accompanying version_info::not_sane_exception
exception type, as well as the version_info::good() method.

The flag is never naturally set to false during version_info
construction/parsing and actual invalid version numbers are always some
variation of "0.0.0foo" once reserialized. The few cases in code where
the flag is forced to be false can be trivially adapted to not require
its existence. In effect, the associated exception was never thrown and
the good() method always returned true, save for the aforementioned
exceptions.

All in all, this was a preposterous waste of time and indentation
levels.