fs: Prevent the BFS version of get_wml_location() from escaping the data dir
This is essentially a port to the BFS implementation of the fs API of
commits 314425ab0e57b32909d324f7d4bf213d62cbd3b5 and
9f458d1bb29cef4318c8511965083edd1128c62e, which were authored separately
from the BFS branch.

Note that this introduces another disparity in behavior between the BFS
and legacy implementations of this API, in that the BFS
get_wml_location() function will now always start at <data dir>/data/
instead of starting at <data dir> in obscure border cases. This is
expected to have no impact on anything at all for reasons discussed on
IRC.