fix assertion failure in tree_view gui2 widget
which could happen when folding a node with a folded subnote in a
treeview with a vertical scrollbar.

get_unfolded_size() is used calculate the size differnce when the node
is (un)folded and should so not return the size 'as if all child nodes
were unfolded' just the unfolded size of the current node.