suppress false positive memory leak reported by valgrind
This is basically the same as a patch I submitted to anura.

The problem is that a standard vector of pointers at static scope
is used to hold a bunch of heap allocated strings. At program
termination these are not deleted so they are reported as a leak.
We add a crude manager object to delete them and suppress the
error reports.