gui2/tscreenshot_notification: Don't obtain file size from caller
The caller's provided size is notoriously unreliable right now --
especially after the addition of PNG support -- because it's obtained
from display::screenshot(), which just does some math voodoo to
calculate the *expected* size of a BMP file and sticks to that.

stat-ing a file is cheap, and simplifies our constructor's signature.
This also fixes PNG files having the wrong size reported.