remove most OpenPandora specific items
The only thing that the ENABLE_PANDORA flag does now is to allow the
resolution of 800x480 and using a special theme automatically.
Everything else is 100% "mainline code". This means if someone wants to
package for pandora several system libraries will have to be
handcompiled and shipped in the pnd including:
* >=boost 1.48 (at least one component (locale) requires this)
* >=SDL_mixer 1.2.12
* VorbisFile