clear the redo stack when toggeling shroud updates
The previous behaviour causes OOS, assume this situation:
Player1s turn, DSU is activated, player1 moves  a unit and then undoes
the move. then he deactivates dsu and redoes the move. The result is
that the move did not uncover shroud (but it should since the move was
now done after DSU was deactivated).