image: Use IMG_Linked_Version() to obtain runtime SDL_image version
This is a function, not a preprocessor macro, so its existence can't be
checked using #ifdef. Furthermore, I can't think of a valid reason for
the SDL devs to arbitrary drop it in a future release without warning,
so we can safely assume it's always present for now.

(Also, it's 'IMG', not 'Image'.)