dont create unit ids based on fake ids.
To do that i replaced the underlying_id  integer with a struct that
holds an integer and has functions to know whether this is fake id or
not.

Fake unit ids can be created via whiteboard or during unsyced contexes
(like wesnoth.sync_choice or ai code)

This commit is intended to fix the case that soneone created a temporary
unit in wesnoth.sync_choice and then returns that units id which might
not be unique   (becasue it is based on a fake id).

This commit also removes a comment that is invalid by now because config
can hold unsigned long long.