Fix unbound memory read (bug #23606)
This was introduced in commit 8e5eb9a8d4edf8da33dd19540495b6aa60b28375,
probably because it's not clear enough in our implementation of MD5 that
the result of MD5::raw_digest() (and util::md5() by proxy) is not a
null-terminated string. I've added a comment to util::md5() to clarify
this in case it comes up again, but what we really should do is replace
the MD5 implementation with something a bit more C++-ish.

Fixed the issue by having the caller request the text representation of
the MD5 digest instead.