add parse method and static name method to MAKE_ENUM
enum.parse("str"); tries convert "str" to an enum and sets enum to "str"
on success. Returns true on success otherwise false. This is useful
because string_to_enum always throws an exception on failure which is
slow if failue is not that unlikeley.

enumtype::name(); returns "enumname" (the name of the type). This is
intended for debugging.