veryfy that synced user commands come from the correct client
Previously it was possible that (modified) clients send user commands
like recruit or move while its not their turn. To prevent this, we now
use the from_side check (that was previously only used for dependent
commands) also for normal synced commands.

The from_side check works like this:
The sending clients adds a from_side= attribute to its commands which
contains which side issued this command.
The server veryfies that that client actually controls that side
The recieveing clients veryfy that this is the side that can currently
issue commands (the currently plaing side).

For compability with older clients  it currently just gives a
ERR_REPLAY. Later it should be changed into a replay::process_error.