Fix another OOS
Previously it could happen that moves of the previous turn were not
sended yet to other players during 'side X turn' and similar events.
This caused OOS because "change_controller_wml" could be sended before
the the moves of the previous turn.

I also removed a virtual froma function that didn't need to be virtual.