add previous_save_id= in [side]
A feature for mp campaigns. previous_save_id lets you specify which user
of teh previous scenario should control this side. For example if
previous_save_id="Konrad" than the user that controlled the side with
the save_id "Konrad" in the previous scenario should control this side.

This also replaces the controller=<number> feature which had similar
intentions but was less flexible.