fix gui2::tlabel
previous labels captured the mouse after clicking on it (I don't know
why). This means that you had to to do another click after selecting
something different. This was specially annyoing an toggle_panels where
you'd often click on a label.

Maybe it was related to link labels, in any case this fixed the much
more common non-link labels.