UtBS: gave new Teaching ability to Garak
The Teaching ability redistributes a unit's experience points to all
the adjacent units on the same side at the start of every turn.
In case that the result of the division isn't integer, the remaining
points are assigned giving priority to the units with lower levels and
less experience.
If no suitable unit is found, the teacher's experience goes back to zero.