gui2/tstacked_widget: Add single layer mode and layer switching
This effectively allows us to use stacked_widget as part of the
implementation of tabbed dialogs like Preferences.

The most important thing here is making sure individual layers are
selected so that they can receive events. This works fine in testing,
and it's much easier for me to take an existing widget and adapt it for
my purposes rather than build a new one from scratch.

Another important thing is not selecting the same layer twice. This
apparently deselects it instead.

Finally, existing applications like the titlescreen are unaffected.