gui2/tgame_version: Turned into a tabbed dialog, added plenty more info
The dialog now includes the game and OS version, and has three tabs with
extra information: the Game Paths tab with the existing contents
inherited from the tgame_paths dialog, the Libraries tab with a table
listing the build/runtime library versions, and the Features tab listing
the status of optional build-time features.

Had to move a lot of code around for this, obviously. Now I have to
decide again where to move the button from Preferences -> General.