gui2/ttitle_screen: Add button to display the About dialog
It's on the bottom left corner, just like the version number. It's not
possible to have both the version number label and the button on the
same layer because only widgets on the topmost stacked_widget layer
receive events, and the label can't be on there because it messes up the
layout of the button panels row.

This currently causes the game to abort on 800x480 because the layout
algorithm runs out of space (needs 800x487) due to to the top row on the
dialog's main grid taking up a fixed amount of vertical space. This will
be corrected in a separate commit.