gui2/ttitle_screen: Reorganize grid slightly
This frees up some space for the central stacked widget so the game
doesn't crash to desktop due to requiring a vertical resolution of 487
or more to fit it with the version label and button.

Also, the layout of the version label and button is now simpler as they
are now on their own grid outside the central stacked widget. This also
happens to ensure that the label doesn't end up behind the tip of the
day panel. Whether this is a good thing or not depends on personal
opinion.

(And incidentally, the button no longer has a margin around it. This is
intentional.)