Add manifest declaring compatibility with Windows 8.1 and 10
This is needed so GetVersionEx doesn't lie to us on these versions.

MSVC++ people: I have no idea how this will interact with your building
procedure, if at all. In particular, I believe MSVC++ is supposed to
produce its own manifest declaring our dependencies on the MSVC++
runtime libraries, and it may or may not emit the compatibility
information as well.