Add correct version numbers to the Windows resource file
For the numeric fields, this means we get to treat all X.Y.Z versions as
X.Y.Z.0, and there's no string suffix for those for obvious reasons. Not
that anyone cares, really.

Now we get a pretty name for our client process in the Task Manager too,
hooray!

(Also cleaned-up some of the placeholder code that was left here by
loonycyborg.)