help: automatically scale unit_type.profile in the absence of unit_type.small_profile
Prior to this, profile= was used in the absence of small_profile, but not scaled, requiring
the manual specification of a 'small_profile=profile~SCALE()' line in unit type files if no
unique small profile image was provided. Since it appears the help browser is the only place
that used small_profile, we move the scaling here.

The standard value of 205x205 is used.