Use listboxs sorting feature in experinemtal gui2 addon_list
Also fixes the toggle panels that show the addons description and the filter
textbox in the non new-widgets codepath.

These changes mostly effect the non new-wigets codepath. The new-widgets
codepath (which already had a sorting feature) still works like before
except that it was adjusted to the changes to the cfg file.