Windows Installer: Translate 'Battle for Wesnoth' and shortcuts
The title is now localized, as are the start menu shortcuts. As localized
titles can be longer than the English string, a third line was explicitly
added to the welcome page (else NSIS cuts the line in half).

The manual shortcut points to the translated manual version, if available.

Due to lack of Unicode support in NSIS 2, only translations matching the
codepage are displayed correctly.

The desktop.ini file isn't removed during uninstallation as it may contain
info on other programs, e.g. if the player chose to place the shortcuts in
'Games'.