Fix the graphics corruption when scrolling with SDL2
SDL2 has issues with overlapping surfaces when blitting, it seems to trigger an internal bug in SDL2. This causes the blitted surface to be incorrectly filled, causing a graphics corruption. This introduces a workaround that does a copy of the original surface before blitting, thus avoiding the bug. This behaviour seems to only occur while scrolling so it's a reasonably rare occurence.