Xcode project: update pango modules
The previous version of the modules worked only with OS X 10.9. This
version works with both 10.10 and 10.9 (and potentially with 10.8; we
currently have no 10.8 system available for testing).

These modules do not work on OS X 10.6 and 10.7. Different methods of
building Wesnoth will be provided for those OS X versions (and
potentially for 10.8).

OS X 10.5 and earlier is no longer supported by Wesnoth 1.13 (that is
unchanged by this commit).