Fix version_info ctor calls expecting the sane parameter
The sane parameter was removed in commit
02a31a5b8945fd6dff235157e317660ae048a89c, but I did not realize that
`false` would still be implicitly cast to char and invoke the usual list
constructor.

Fix this by using the default constructor instead, which does what we
want sans removed parameter.

Note that this mistake and subsequent fix has no bearing on anything
other than code presentation because false and '\0' are the same value
and the three number values are 0 with the default ctor too.