Fix alpha blending issue when scrolling under SDL2
SDL2 has a workaround for a broken blitting operation, and on some
hardware accelerated systems this led to an alpha blending issue where
map regions that were frequently redrawn would grow progressively
darker until the alpha blended aspects are black. This disables
alpha-blending on the copy of the scrolled map area as no further
alpha composition should be done without a complete redraw of the
affected area.