Resolve colour cursor blending issue in SDL 2.0
With the addition of surface blending in SDL2, having colour cursors enabled results in a faint trail of mouse cursors being left behind. It is stronger when the mouse moves more slowly and definitely noticeable when the cursor is locked for the adding new hot-key function.

This is related to https://github.com/wesnoth/wesnoth/pull/473 but unlike that issue, this was a problem for me in Windows. I do not know if it is also an issue with Linux and OS X builds.