Show the non-rebindable shortcuts in preferences.
They can't be changed, but it will at least let people know that the key works for that command.
Currently they're just hard-coded. A more general solution would be possible, if there's
any chance that more non-rebindable hotkeys are needed.

This also removes Cmd+Shift+Q as a shortcut for Quit to Desktop on OSX.