refactor playsingle_controller::before_human_turn
Also removed a raise_turn_started event at the beginning of human turns
which didn't fire at the beginning of networked turns.

I don't know what the raise_turn_started did if this actually breaks
anything feel free to read it but please add a comment on why it fires
on human turns but not on networked turns.