Resolve Bug #20036
When a map is generated, the labels for the player starting locations are correctly set. However, generating another map (or even loading a map and then generating a new one) results in the old labels remaining even after the new map has been generated and displayed.

Issue appears to be that the reset labels flag is not being set after a map generation. This change sets the reset labels flag and the subsequent perform_refresh() function then includes a reset of the map labels.