Replace 255 with SDL_ALPHA_OPAQUE where appropriate.
This is an extension of the previous correction to alpha values in attack_prediction_display.cpp.
For clarity and compatibility, use the SDL_ALPHA_OPAQUE definition instead of assuming hard-coded 255. This is defined in both SDL 1.2 and 2.0.
Where there are mathematical operations on alpha channel which assumes a value of 255, leave alpha as hard-coded 255.
Old versions of SDL treated 255 as transparent and 0 as opaque - there are still instances where alpha is 0 in a non-transparency situation, but these seem okay. Possibly because there is no alpha blending involved.