Realign translated logo images.
Ever since the logo was split into separate image (swords and shield) and text, as well as the dimensions being modified, translated logos have been misaligned.
While it's true the translated logo images should be remade so that it the image and text are similarly separated, we can still align the existing logos with the current logo background image.

In many cases, the translated logos were not even showing on the loading screen because its dimensions broke a check in blit_modification::operator() (translated logo height > background logo height).
After this change, all translated logos should display again (and are even nicely aligned on top of the current logo background image).