add enable_if= to mod and era events
http://gna.org/bugs/?23721
enable_if = in [event] contains a condition that is evalulated before
the scenario is startd (specially before the [replay_start] snapshot is
taken) to check whether the event should be added to the scenario.
[event]
enable_if="$($scenario_num -1 > -1)"
[/event]
will for example add the event only if it's not the first scenario (assuming
wml sets the variable) 'scenario_num' to the current scenario number.