Updated embedded Lua library to 5.3.1
This constitutes a copy of the src directory of the 5.3.1 source, with extension
changes (.c -> .cpp). Applicable function names were changed.

Additionally, luaL_typeerror was manually readded, as this  function is no longer
included in the source as of 5.2.

The changes from the following commits were manually reapplied:
  * 6655b186b1bd5301f7171cd13d3307af3f700bd7
    LUAI_EXTRASPACE

The following patches were applied:
  * http://www.lua.org/bugs.html#5.3.1
    io.lines does not check maximum number of options.